Restructuring and adjusting the types a bit
