+++ .prettierignore
@@ -1,0 +1,5 @@
+.turbo
+coverage
+dist
+package.json
+yarn.lock
