+++ .yarnrc.yml
@@ -1,1 +1,2 @@
 yarnPath: .yarn/releases/yarn-4.0.2.cjs
+nodeLinker: node-modules
