+++ lint-staged.config.mjs
@@ -1,0 +1,8 @@
+export default {
+  "**/*.{ts,js,json,md,yml}": [
+    "prettier --write"
+  ],
+  "**/*.{ts,js}": [
+    "eslint --fix"
+  ]
+}
