+++ package.json
@@ -1,54 +1,74 @@
 {
   "name": "async-sema",
-  "version": "3.1.2",
+  "version": "4.0.0",
   "description": "Semaphore using `async` and `await`",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/vercel/async-sema.git"
+    "url": "git+https://github.com/jdalrymple/async-sema.git"
   },
-  "author": "Olli Vanhoja",
+  "author": "Justin Dalrymple",
   "keywords": [
     "semaphore",
     "async",
     "await"
   ],
-  "homepage": "https://github.com/vercel/async-sema",
+  "homepage": "https://github.com/jdalrymple/async-sema",
   "bugs": {
-    "url": "https://github.com/vercel/async-sema/issues"
+    "url": "https://github.com/jdalrymple/async-sema/issues"
   },
   "license": "MIT",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.js"
+    }
+  },
   "files": [
-    "lib"
+    "dist"
   ],
-  "main": "lib/index.js",
   "scripts": {
-    "build": "tsc",
-    "lint:staged": "lint-staged",
-    "prepublishOnly": "yarn build",
-    "prettier": "prettier --write --single-quote './{src,test}/**/*.ts'",
-    "test": "jest"
+    "build": "tsup src/index.ts --format esm,cjs --dts --treeshake",
+    "test:types": "tsc",
+    "test:unit": "jest --maxWorkers=50% test",
+    "format:docs": "prettier '**/(*.json|.yml|.js|.md)' --ignore-path ./.prettierignore",
+    "format:docs:fix": "yarn format:docs --write",
+    "format:src": "prettier '{src,test}/**/*.ts' --ignore-path ./.prettierignore",
+    "format:src:fix": "yarn format:src --write",
+    "format": "yarn format:src && yarn format:docs",
+    "format:fix": "yarn format:src:fix && yarn format:docs:fix",
+    "lint": "eslint '{src,test,scripts}/**/*.ts'",
+    "lint:fix": "yarn lint --fix",
+    "release:rc": "auto next",
+    "release:canary": "auto canary",
+    "release": "auto shipit",
+    "postinstall": "husky install",
+    "prepublishOnly": "pinst --disable",
+    "postpublish": "pinst --enable"
   },
   "devDependencies": {
+    "@swc/core": "^1.3.96",
+    "@swc/jest": "^0.2.29",
     "@types/jest": "^29.5.8",
-    "@types/node": "^20.9.1",
+    "@types/node": "^20.9.2",
+    "@typescript-eslint/eslint-plugin": "^6.11.0",
+    "@typescript-eslint/parser": "^6.11.0",
+    "auto": "^11.0.4",
+    "eslint": "^8.54.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jest": "^27.6.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "husky": ">=8.0.3",
     "jest": "^29.7.0",
-    "lint-staged": "^15.1.0",
-    "pre-commit": "^1.2.2",
+    "jest-junit": "^16.0.0",
+    "lint-staged": ">=15.1.0",
+    "pinst": ">=3.0.0",
     "prettier": "^3.1.0",
-    "ts-jest": "^29.1.1",
+    "tsup": "^8.0.0",
     "typescript": "^5.2.2"
-  },
-  "pre-commit": "lint:staged",
-  "lint-staged": {
-    "*.{js,ts}": [
-      "prettier --write --single-quote",
-      "git add"
-    ]
-  },
-  "jest": {
-    "preset": "ts-jest",
-    "verbose": false,
-    "testRegex": "test\\/.*\\.ts$"
   },
   "packageManager": "yarn@4.0.2"
 }
